Broken down home view into individual components
