+++ src/App.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Grid } from "react-bootstrap";
 import { Route } from "react-router-dom";
 
-import Home from "./Home";
+import Home from "./Home/Home";
 import Navigation from "./Navigation";
 import Footer from "./Footer";
